src/api/requirements.txt
flask
requests
python-dotenv==1.0.1
gunicorn==21.2.0
flask-cors==4.0.0
whl/prompty-0.1.20240317.171740-py3-none-any.whl